Fix EnableRequestInterception
In recent versions of Chrome Network.enableRequestInterception has been
changed to Network.setRequestInterception. It is now also possible to
specify options for which requests should be intercepted. I'm not show
how you feel about breaking your API and changing the parameters to
EnableRequestInterception or maybe it should be a new function?